requirements.txt
Django>=4.2,<5.0
pillow
gunicorn
whitenoise
django-crispy-forms
python-dotenv
sendgrid
pdfminer.six
google-genai
PyPDF2 
python-docx